+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Add missing `paymentMethodToken` search to `CreditCardVerificationSearchRequest`
+
 ## 2.108.0
 * Fix null pointer exception when 3DS lookup returns a `422`
 * Add ThreeDSecure test payment method nonces